ep：封装添加试卷参数接口api

+++ src/main/java/com/csswust/patest2/controller/ep/EpExamParamAction.java
@@ -1,12 +1,11 @@
-package com.csswust.patest2.controller.lexam;
+package com.csswust.patest2.controller.ep;
 
 import com.csswust.patest2.common.APIResult;
 import com.csswust.patest2.controller.common.BaseAction;
-import com.csswust.patest2.dao.CourseInfoDao;
 import com.csswust.patest2.dao.ExamParamDao;
-import com.csswust.patest2.dao.KnowledgeInfoDao;
 import com.csswust.patest2.entity.ExamParam;
 import com.csswust.patest2.service.ExamParamService;
+import com.csswust.patest2.service.common.AuthService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -19,19 +18,27 @@ import org.springframework.web.bind.annotation.RestController;
  * Created by 972536780 on 2018/3/19.
  */
 @RestController
-@RequestMapping("/examParam")
-public class ExamParamAction extends BaseAction {
-    private static Logger log = LoggerFactory.getLogger(ExamParamAction.class);
+@RequestMapping("/ep/examParam")
+public class EpExamParamAction extends BaseAction {
+    private static Logger log = LoggerFactory.getLogger(EpExamParamAction.class);
 
     @Autowired
     private ExamParamService examParamService;
+    @Autowired
+    private AuthService authService;
+    @Autowired
+    private ExamParamDao examParamDao;
 
     @RequestMapping(value = "/selectByCondition", method = {RequestMethod.GET, RequestMethod.POST})
     public Object selectByCondition(
-            ExamParam examParam,
+            @RequestParam Integer examId,
             @RequestParam(required = false) Integer page,
             @RequestParam(required = false) Integer rows) {
-        if (examParam == null) return new APIResult(-501, "examParam不能为空");
+        if (!authService.judgeEpAuth(getEpUserId(), examId)) {
+            return new APIResult(-501, "权限不足");
+        }
+        ExamParam examParam = new ExamParam();
+        examParam.setExamId(examId);
         return examParamService.selectByCondition(examParam, page, rows);
     }
 
@@ -40,7 +47,9 @@ public class ExamParamAction extends BaseAction {
             @RequestParam(required = false) Integer knowId,
             @RequestParam(required = false) Integer levelId,
             @RequestParam Integer examId) {
-        if (examId == null) new APIResult(-501, "examId不能为空");
+        if (!authService.judgeEpAuth(getEpUserId(), examId)) {
+            return new APIResult(-501, "权限不足");
+        }
         return examParamService.selectProblemTotal(knowId, levelId, examId);
     }
 
@@ -50,17 +59,19 @@ public class ExamParamAction extends BaseAction {
             @RequestParam Integer[] knowIds,
             @RequestParam Integer[] levels,
             @RequestParam Integer[] scores) {
-        if (examId == null) return new APIResult(-501, "examId不能为空");
+        if (!authService.judgeEpAuth(getEpUserId(), examId)) {
+            return new APIResult(-501, "权限不足");
+        }
         return examParamService.insertByArray(examId, knowIds, levels, scores);
-    }
-
-    @RequestMapping(value = "/deleteByIds", method = {RequestMethod.GET, RequestMethod.POST})
-    public Object deleteByIds(@RequestParam String ids) {
-        return examParamService.deleteByIds(ids);
     }
 
     @RequestMapping(value = "/deleteById", method = {RequestMethod.GET, RequestMethod.POST})
     public Object deleteById(@RequestParam Integer id) {
+        ExamParam examParam = examParamDao.selectByPrimaryKey(id);
+        if (examParam == null) return null;
+        if (!authService.judgeEpAuth(getEpUserId(), examParam.getExamId())) {
+            return new APIResult(-501, "权限不足");
+        }
         return examParamService.deleteById(id);
     }
 }
